fix conversion warnings

+++ src/core.h
@@ -225,9 +225,9 @@ void fill_segment(const argon2_instance_t *instance,
                   argon2_position_t position);
 
 // related to precomputations...
-uint64_t argon2i_index_size(const argon2_instance_t *instance);
+uint32_t argon2i_index_size(const argon2_instance_t *instance);
 
-uint64_t argon2i_precompute(const argon2_instance_t *instance,
+uint32_t argon2i_precompute(const argon2_instance_t *instance,
                             argon2_precomputed_index_t *oIndex);
 
 void fill_memory_blocks_precompute(const argon2_instance_t *instance);
